+++ tt_metal/tt_metal.cpp
@@ -314,7 +314,7 @@ inline void SetRuntimeArgsImpl(
 
 namespace detail {
 
-bool WriteToDeviceDRAMChannel(IDevice* device, int dram_channel, uint64_t address, std::vector<uint32_t>& host_buffer) {
+bool WriteToDeviceDRAMChannel(IDevice* device, int dram_channel, uint32_t address, std::vector<uint32_t>& host_buffer) {
     bool pass = true;
     TT_FATAL(
         address >= device->allocator()->get_base_allocator_addr(HalMemType::DRAM),
@@ -326,7 +326,7 @@ bool WriteToDeviceDRAMChannel(IDevice* device, int dram_channel, uint64_t addres
 }
 
 bool ReadFromDeviceDRAMChannel(
-    IDevice* device, int dram_channel, uint64_t address, uint32_t size, std::vector<uint32_t>& host_buffer) {
+    IDevice* device, int dram_channel, uint32_t address, uint32_t size, std::vector<uint32_t>& host_buffer) {
     bool pass = true;
     tt::tt_metal::MetalContext::instance().get_cluster().dram_barrier(device->id());
     host_buffer.resize((size + sizeof(uint32_t) - 1) / sizeof(uint32_t));
